requirements.txt
pyyaml
openai>=1.0.0
pytest
coverage
